cdk-sih/requirements.txt
aws-cdk-lib==2.165.0
constructs>=10.0.0,<11.0.0
netaddr~=0.8.0
python-dotenv~=0.19.2
pytz~=2021.3
PyYAML~=6.0
